AttackModule::size returns f32, not u64

+++ src/cpp.rs
@@ -5932,7 +5932,7 @@ pub mod root {
                     pub fn size(
                         arg1: *mut root::app::BattleObjectModuleAccessor,
                         arg2: libc::c_int,
-                    ) -> u64;
+                    ) -> f32;
                 }
                 extern "C" {
                     #[link_name = "\u{1}_ZN3app8lua_bind27AttackModule__set_size_implEPNS_26BattleObjectModuleAccessorEif"]
